src/utils/gemini_fallback.py
"""
Gemini AI Fallback Recognition Module
=====================================

This module provides fallback character and expression recognition capabilities
using Google's Gemini AI when local OCR models have low confidence or fail.

Features:
- Character recognition for unclear handwritten symbols
- Full mathematical expression recognition
- Step-by-step equation solving
- Intelligent fallback based on confidence thresholds
- Rate limiting and retry logic for API reliability
"""

import os
import json
import time
import logging
from typing import Dict, Any, Optional, List
from dataclasses import dataclass
import cv2
import numpy as np
from PIL import Image

# Configure logging
logger = logging.getLogger(__name__)

@dataclass
class GeminiConfig:
    """Configuration for Gemini API integration"""
    api_key: str
    model_name: str = "gemini-1.5-pro"
    max_tokens: int = 1024
    temperature: float = 0.1
    rate_limit_delay: float = 1.0
    max_retries: int = 3
    timeout: int = 30

class GeminiFallbackRecognizer:
    """
    Fallback recognizer using Google Gemini API for difficult character recognition
    """
    
    def __init__(self, config: GeminiConfig):
        """
        Initialize Gemini fallback recognizer
        
        Args:
            config: Configuration object with API settings
        """
        self.config = config
        self.model = None
        self.last_request_time = 0
        
        # Try to initialize Gemini (will be None if not available)
        try:
            import google.generativeai as genai
            genai.configure(api_key=config.api_key)
            self.model = genai.GenerativeModel(config.model_name)
            self.genai = genai
            logger.info(f"Initialized Gemini model: {config.model_name}")
        except ImportError:
            logger.warning("Google GenerativeAI not available - install with: pip install google-generativeai")
        except Exception as e:
            logger.warning(f"Failed to initialize Gemini API: {e}")
    
    def is_available(self) -> bool:
        """Check if Gemini API is available and configured"""
        return self.model is not None
    
    @classmethod
    def from_env(cls, model_name: str = "gemini-1.5-pro") -> 'GeminiFallbackRecognizer':
        """
        Create GeminiFallbackRecognizer from environment variables
        
        Args:
            model_name: Gemini model to use
            
        Returns:
            GeminiFallbackRecognizer instance
        """
        api_key = os.getenv('GEMINI_API_KEY')
        if not api_key:
            raise ValueError("GEMINI_API_KEY environment variable not set")
        
        config = GeminiConfig(
            api_key=api_key,
            model_name=model_name
        )
        
        return cls(config)


# Utility functions for integration with main system

def create_fallback_recognizer() -> Optional[GeminiFallbackRecognizer]:
    """
    Create fallback recognizer if API key is available
    
    Returns:
        GeminiFallbackRecognizer instance or None if not configured
    """
    try:
        return GeminiFallbackRecognizer.from_env()
    except Exception as e:
        logger.warning(f"Could not create Gemini fallback recognizer: {e}")
        return None

def should_use_fallback(confidence: float, threshold: float = 0.7) -> bool:
    """
    Determine if fallback should be used based on confidence
    
    Args:
        confidence: Confidence score from local model
        threshold: Minimum confidence threshold
        
    Returns:
        True if fallback should be used
    """
    return confidence < threshold
