requirements.txt
# Core Dependencies
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.3.0
scipy>=1.7.0
matplotlib>=3.4.0

# Machine Learning & Deep Learning
torch>=1.9.0
torchvision>=0.10.0
scikit-image>=0.18.0
albumentations>=1.3.0

# Symbolic Mathematics
sympy>=1.9.0

# Web Framework & API
flask>=2.0.0
werkzeug>=2.0.0

# Google AI Integration (Optional)
google-generativeai>=0.3.0

# Image Processing & Computer Vision
opencv-python-headless>=4.5.0

# Data Processing
pandas>=1.3.0
pyyaml>=6.0

# Development & Testing
tqdm>=4.62.0
pytest>=6.2.0

# Optional Dependencies for Enhanced Features
# tensorboard>=2.7.0  # For training visualization
# jupyter>=1.0.0      # For notebook development
# streamlit>=1.0.0    # Alternative web interface
